+++ learn_project/__init__.py
@@ -1,11 +1,12 @@
 #запуск сервера
 #set FLASK_APP=webapp && set FLASK_ENV=development && 
 #set FLASK_DEBUG=1 && flask run
-from flask import Flask, render_template, request, url_for
-from flask_login import LoginManager, current_user
-from learn_project.model import db, Products, Images
+from flask import Flask, render_template
+from flask_login import LoginManager
+from learn_project.model import db
 from learn_project.user.model import Users
 from learn_project.user.views import blueprint as user_blueprint
+from learn_project.advert.views import blueprint as advert_blueprint
 from flask_migrate import Migrate
 
 
@@ -17,6 +18,7 @@ def create_app():
     login_manager.init_app(app)
     login_manager.login_view = 'user.login'
     app.register_blueprint(user_blueprint)
+    app.register_blueprint(advert_blueprint)
     migrate = Migrate(app, db)
 
     @login_manager.user_loader
@@ -25,33 +27,6 @@ def create_app():
 
     @app.route('/')  					 # путь, перейдя по котрому запустится app
     def index():  						 # возвращает стартовую страничку
-        return render_template('adverts/index.html', page_title='Тут вам не авито!')
-
-    @app.route('/products')
-    def products():
-
-        title = 'Товары'
-        page = request.args.get('page', 1, type=int)
-        products_list = Products.query.order_by(Products.date.desc()).paginate(page,
-                                                                               app.config['ITEMS_PER_PAGE'],
-                                                                               False)
-        prev_url = url_for('products', page=products_list.prev_num) \
-            if products_list.has_prev else None
-
-        next_url = url_for('products', page=products_list.next_num) \
-            if products_list.has_next else None
-
-        return render_template('adverts/products.html',
-                                page_title=title,
-                                products_list=products_list.items,
-                                next_url=next_url,
-                                prev_url=prev_url)
-
-    @app.route('/ad_page/<prod_db_id>')
-    def ad_page(prod_db_id):
-        ad_items = Products.query.get(prod_db_id)
-        image_urls = Images.query.filter_by(product_id=prod_db_id).all()  # возвращает список объектов класса
-        image_urls = [image_url.img_url for image_url in image_urls]      # вытаскиваем из этих объектов ссылки и кладём в список
-        return render_template('adverts/ad_page.html', ad_items=ad_items, image_urls=image_urls)
+        return render_template('advert/index.html', page_title='Тут вам не авито!')
 
     return app  # возвращает экземпляр приложения Flask
