+++ learn_project/templates/advert/index.html
@@ -14,4 +14,4 @@
         {% endwith %}
       </div>
   </div>
-{% endblock %}   
+{% endblock %}
