--- learn_project/templates/adverts/ad_page.html
@@ -1,53 +1,0 @@
-<!doctype html>
-<html lang="ru">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-
-    <title>Learn Project!</title>
-  </head>
-  <body>
-    <div class="container">
-    <h1>Тут вам не авито</h1>
-    <div class="row">
-      <div class="col-8">
-        <div class="card mt-4">
-          <div class="card-body">
-            <h3 class="card-title">{{ad_items.name}}</h3>
-            <h4>{{ad_items.price}} р. </h4>
-            <p class="card-text">{{ad_items.text}}</p>
-            <small class="text-muted">{{ad_items.date}}</small>
-          </div>
-        </div>
-      </div>
-      <div class="col-4"> 
-        <div class="card mt-4">
-          <div class="card-body">
-            {% for url in image_urls %}
-            <img src="{{url}}" width="100" height="75">
-            {% endfor %}
-        </div>
-        </div>
-      </div>
-    </div>
-    <div class="card card-outline-secondary my-4">
-      <div class="card-header">
-        {{ad_items.ad_number}}
-      </div>
-      <div class="card-body">
-        <p>{{ad_items.address}}</p>
-      </div>
-    </div>
-    </div>
-    
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js" integrity="sha384-OgVRvuATP1z7JjHLkuOU7Xw704+h835Lr+6QL9UvYjZE3Ipu6Tp75j7Bh/kR0JKI" crossorigin="anonymous"></script>
-  </body>
-</html>
